Added a maven pom.xml file that employs an alternative method of joint compiling the Java and groovy sources using a custom build of the groovy-eclipse-compiler maven plugin, which successfully compiles the groovy serializer where other methods that are more mainstream such as gmavenplus failed. The groovy docs do mention that this is the only method that will offer a stub-less joint compile of Groovy/Java, although I had to update the groovy-eclipse repo to build against groovy 2.4.6 as they seem to be on 2.4.3 right now - might have been unecessary...
This is the only way I have been able to get CompactBufferSerializer.groovy
to compile after trying GMavenPlus versions 1.0-1.5 and 1.6-SNAPSHOT, Groovy 2.5.0-SNAPSHOT
built from the repo and all subproject jars manually installed vs. 2.4.6 as a dependency,
and several different ways of converying the intent in CompactBufferSerializer.groovy in
order to get a stub to generate properly as various GROOVY/GRECLIPSE issues and SO
questions led me to believe this could make a difference. The main issue was that the
stub just didn't carry forward the type information conveyed by the method signature
in the groovy source and/or the compiler (groovyc or javac?) didn't understand how to
interpret the SerializerShim<T> interface. Some clues as to what is going on here were
found in the groovy-eclipse project commit that updated to groovy 2.4.3 from 2.3.x, if
I was really on top of it I would list the commit id and those GROOVY-XXXX issues here
but I will abstain due to router issues preventing my accessing github.com right now.
This breadcrumb will have to suffice.

In summary, this was required:
1) git clone groovy-eclipse
2) download Eclipse SDK 4.4
3) update various jars, configurations, and other files for groovy 2.4.3 -> 2.4.6 support.
   - copied off of the 2.3.7 -> 2.4.3 commit already present in the groovy-eclipse project.
4) cd into the groovy-eclipse/extras/groovy-eclipse-batch-builder directory
5) follow the directions in the comments in the ./build.xml file
   - mostly editing ./build.properties with paths
6) run `ant extract-create-install`
   - should fail over and over saying "there is no blah/blah/.../blah/bin directory!"
   - just mkdir the missing directory and keep going
7) if you didn't edit jar.dir=/tmp/uberjar/lib in build.properties, look there for the jar
   - even though we failed installing it to the local m2 repo, so fix that with mvn install:install-file ...
8) now cd into the groovy-eclipse/extras/groovy-eclipse-compiler, edit the pom file
   - just point to the artifact you manually installed above
   - disable the gpg plugin by commenting it out
9) run `mvn clean install`
10) in your pom.xml set <configuration><compilerId>groovy-eclipse-compiler... for maven-compiler-plugin version 3.0
    - presumably this is the project that has your Groovy serializer extending SerializerShim<T>
    - include the two artifacts you just installed above as dependencies here.
11) mvn clean install!

Now to see if this will acutally work we we try it with Spark. Hope so.